Modified folder structure. Made an attempt and making the main page look pretty. Could do with a Designer really...
