+++ app/browserWindows/main/main.css
@@ -1,0 +1,4 @@
+body {
+    max-width: 700px;
+    margin: 0 auto;
+}
