--- app/index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8">
-    <title>Dotnet Runner</title>    
-    <script src="rendererMain.js"></script>
-    <script src="index.js"></script>
-    <style>
-      body {
-        font-family: -apple-system,BlinkMacSystemFont,Segoe WPC,Segoe UI,HelveticaNeue-Light,Ubuntu,Droid Sans,sans-serif;
-        line-height: 22px;
-      }
-    </style>
-  </head>
-  <body>
-        <button class="purge">Purge (<span class="runningProcessCount">0</span>)</button>
-        <button class="start-all">Start all</button>
-        <button class="terminate-all">Terminate all</button>
-        <div class="runners"></div>
-  </body>
-</html>
