--- app/states.js
@@ -1,6 +1,0 @@
-module.exports = {
-    stopped: 0,
-    starting: 1,
-    running: 2,
-    stopping: 3
-};
