+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "dotnet-runner",
-  "version": "0.0.2",
+  "version": "0.0.3",
   "description": "Electron application to launch dotnet applications",
   "main": "startup.js",
   "scripts": {
